Fix tests  - Split redux from ServiceGraphPage to only test the non redux part
