+++ src/pages/ServiceGraph/ServiceGraphRouteHandler.tsx
@@ -4,7 +4,7 @@ import { PropTypes } from 'prop-types';
 
 import { GraphParamsType } from '../../types/Graph';
 import * as LayoutDictionary from '../../components/CytoscapeLayout/graphs/LayoutDictionary';
-import ServiceGraphPage from './ServiceGraphPage';
+import ServiceGraphPage from './ServiceGraphPageContainer';
 
 const URLSearchParams = require('url-search-params');
